Move user and management user model to user folder
